IN-41 refactor: create temp for renaming
