+++ src/shared/ui/ButtonTemp/Button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { Button } from '.'
+import { Button } from './Button'
 
 const meta = {
   argTypes: {
