+++ src/shared/ui/CheckboxTemp/Checkbox.module.scss
