+++ src/shared/ui/CheckboxTemp/checkBoxSelectedIcon.tsx
