+++ src/shared/ui/DialogTemp/Dialog.module.scss
