+++ src/shared/ui/DialogTemp/Dialog.stories.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react'
 import { action } from '@storybook/addon-actions'
 import { Meta, StoryObj } from '@storybook/react'
 
-import { Button, Typography } from '../'
-import { Dialog } from './dialog'
+import { Button, Typography } from '..'
+import { Dialog } from './Dialog'
 
 const meta = {
   argTypes: {
