+++ src/shared/ui/DialogTemp/Dialog.tsx
@@ -3,7 +3,7 @@ import { Fragment } from 'react'
 import { Button, Modal, ModalProps } from '@/src/shared/ui'
 import { clsx } from 'clsx'
 
-import s from './dialog.module.scss'
+import s from './Dialog.module.scss'
 
 type Props = {
   buttonsContainerClassName?: string
