+++ src/shared/ui/Dropdown/index.ts
@@ -1,0 +1,3 @@
+export * from './DropdownMenu/DropdownMenu'
+export * from './DropdownItem/DropdownItem'
+export * from './DropdownSeparator/DropdownSeparator'
