+++ src/shared/ui/ImageUploadInputTemp/ImageUploadInput.stories.tsx
