+++ src/shared/ui/ModalTemp/Modal.module.scss
