+++ src/shared/ui/ModalTemp/Modal.tsx
@@ -11,10 +11,10 @@ import {
 import { clsx } from 'clsx'
 import { AnimatePresence, motion } from 'framer-motion'
 
-import s from './modal.module.scss'
+import s from './Modal.module.scss'
 
 import { ArrowIosBack, Close } from '../../assets/icons'
-import { Button } from '../button'
+import { Button } from '../ButtonTemp/Button'
 
 export type ModalSize = 'lg' | 'md' | 'sm'
