+++ src/shared/ui/ScrollAreaTemp/ScrollArea.module.scss
