+++ src/shared/ui/SpinnerTemp/Spinner.module.scss
