+++ src/shared/ui/SpinnerTemp/Spinner.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { Spinner } from '.'
+import { Spinner } from './Spinner'
 
 const meta = {
   component: Spinner,
