+++ src/shared/ui/SpinnerTemp/Spinner.tsx
@@ -2,7 +2,7 @@ import { ComponentProps, FC } from 'react'
 
 import { clsx } from 'clsx'
 
-import s from './spinner.module.scss'
+import s from './Spinner.module.scss'
 
 export const Spinner: FC<ComponentProps<'div'>> = ({
   className,
