+++ src/shared/ui/TypographyTemp/Typography.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { Typography } from './index'
+import { Typography } from './Typography'
 import { TypographyVariants } from './typographyVariants'
 
 const meta = {
