--- src/shared/ui/dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './dialog'
