+++ src/shared/ui/index.ts
@@ -1,22 +1,20 @@
-export * from './Checkbox'
-export * from './Controlled'
-export * from './TextField'
-export * from './Typography'
-export * from './button'
-export * from './Card/Card'
-export * from './dialog'
-export * from './modal'
-export * from './scrollArea'
-export * from './select'
-export * from './Dropdown/DropdownItem/DropdownItem'
-export * from './Dropdown/DropdownMenu/DropdownMenu'
-export * from './Dropdown/DropdownSeparator/DropdownSeparator'
-export * from './spinner'
-export * from './DatePicker/DatePicker'
 export * from './Avatar/Avatar'
-export * from './TextArea/TextArea'
-export * from './ImageUploader/imageUploader'
-export * from './ImageUploadInput/imageUploadInput'
-export * from './Tabs'
+export * from './ButtonTemp/Button'
+export * from './Card/Card'
+export * from './CheckboxTemp/Checkbox'
 export * from './Combobox/Combobox'
+export * from './Controlled'
+export * from './DatePicker/DatePicker'
+export * from './DialogTemp/Dialog'
+export * from './Dropdown'
+export * from './ImageUploaderTemp/ImageUploader'
+export * from './ImageUploadInputTemp/ImageUploadInput'
+export * from './ModalTemp/Modal'
 export * from './Popover/Popover'
+export * from './ScrollAreaTemp/ScrollArea'
+export * from './SelectTemp/Select'
+export * from './SpinnerTemp/Spinner'
+export * from './Tabs'
+export * from './TextArea/TextArea'
+export * from './TextFieldTemp/TextField'
+export * from './TypographyTemp/Typography'
